+++ src/AxeFxControl.h
@@ -4,4 +4,4 @@
 // #define AXE_DEBUG_SYSEX
 #define DEBUGGER 								Serial
 
-#include "AxeSystem.h" 
+#include "interface/AxeSystem.h"
