+++ src/interface/AxeEffect.h
@@ -1,9 +1,9 @@
 #pragma once
 
 #include <Arduino.h>
-#include "AxeTypes.h"
-#include "AxeEffectEnum.h"
-#include "AxeEffectLabels.h"
+#include "private/AxeTypes.h"
+#include "private/AxeEffectEnum.h"
+#include "private/AxeEffectLabels.h"
 
 // You can retrieve a list of effects from an AxePreset object.
 class AxeEffect {
@@ -59,39 +59,10 @@ class AxeEffect {
 		// hacking your own sysex that wasn't published in the spec.
 		EffectId getEffectId() { return _effectId; }
 
-	public:
-
 		const static byte MAX_EFFECT_NAME	= 20;
 		const static byte MAX_EFFECT_TAG	= 5;
 
-	protected:
-
-		friend class AxePreset;
-		friend class AxeSystem;
-
-		AxeEffect() {}
-
-		void setAxeSystem(AxeSystem *axe) { _axe = axe; }
-		void setBypassed(bool bypassed) { _bypassed = bypassed; }
-		void setChannelCount(byte count) { _numChannels = count; }
-		void setChannel(Channel channel) { _channel = channel; }
-		void setEffectId(EffectId effectId) { _effectId = effectId; }
-
 	private:
-
-		const static EFFECT_ID_ FIRST_EFFECT = ID_COMP1;
-		const static char ASCII_ZERO = 0x30;
-
-		EffectId _effectId;
-		Channel _channel;
-		byte _numChannels;
-		bool _bypassed;
-
-		byte offset();
-		byte hash();
-		byte blockNum();
-		char blockNumChar();
-
-		AxeSystem *_axe = nullptr;
+		#include "private/AxeEffect_Private.h"
 
 };
