+++ src/interface/AxePreset.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <Arduino.h>
-#include "AxeTypes.h"
+#include "private/AxeTypes.h"
 #include "AxeEffect.h"
 
 // This class represents the current preset. You can get it by calling getPreset() on AxeSystem,
@@ -36,41 +36,15 @@ class AxePreset {
 		byte getMaxEffects() { return _maxEffects; }
 		void setMaxEffects(byte max) { _maxEffects = max; }
 
-	protected:
-		
-		friend class AxeSystem;
-		AxePreset() {}
+		//Our internal limit
+		const static byte MAX_EFFECTS			= 50;
 
-		// These methods are used by AxeSystem to build the preset.
-		// You don't need to call them, and you can't.
-		void setPresetNumber(int number) {  _preset = number; }
-		void setSceneNumber(int number) {  _scene = number; }
-		void setPresetName(char *buffer);
-		void setSceneName(char *buffer);
-		void reset();	
-		bool isComplete();
-
-	public:
-	
 		// These maxiumums are set by the Axe.
 		const static byte MAX_PRESET_NAME	= 32;
 		const static byte MAX_SCENE_NAME	= 32;
 
-		//Our internal limit
-		const static byte MAX_EFFECTS			= 50;
-
 	private:
-
-		void setEffects(const AxeEffect effects[], const unsigned count);
-
-		int _scene = -1;
-		int _preset = -1;
-		int _effectCount = -1;	
-		byte _maxEffects = MAX_EFFECTS;
-		bool _presetNameReceived = false, _sceneNameReceived = false;
-		char _presetName[MAX_PRESET_NAME + 1];
-		char _sceneName[MAX_SCENE_NAME + 1];
-		AxeEffect _effects[MAX_EFFECTS];
+		#include "private/AxePreset_Private.h"		
 		
 };
