+++ src/interface/private/AxeSystem_Midi.cpp
@@ -1,4 +1,4 @@
-#include "AxeSystem.h"
+#include "interface/AxeSystem.h"
 
 #define SER_AVLB _serial->available()
 #define SER_READ _serial->read()
